Update readme.md for correctness and formatting
The current implementation of saveImage does not just take a `figure`
from `getFigure`.  It expects an associative array where one of the
elements is a `figure` and the other elements specify image format,
width, and height.  Updated the documentation to reflect this.  The
right fix would be to take in a `figure` to `saveImage` along with an
options object specifying format, width, and height, to make the API
easier to understand. That would be a breaking change, though, so will
need discussion first.